Create README.mdMaking away for a better tomorrow..
Life has been difficult this much I know it should be easier. What is money to you it holds no value and holds no soul and holds no life but it's the same thing that you kill this tool and destroy for you work hard and nothing get to enjoy your life and it traps you and it makes you something other than what you should be we have to start off somewhere. If everybody give to one another this world will go without want or need if we were capable to give and we all should be this world will go without greed. Behind the back of marriages in front of faces of the children of God are cheated out there inheritance their entitlement agree blinds them for they only want to believe in a little bit but not a lot people will use apps to cover their true faces. Which is wizards and warlocks and the sorcery goes deep and the roots of people without receiving to the veil is covering their face they are blind and hearing aid they do not hear here is my ears but my heart is a ear for listening I speak to all the children of God is a voice of God will touch your hearts and he will take the veil off your hearts he may lead you in the Great Awakening the time is now.